Added new transitions package (unintegrated)
This new package provides a API and a callback mechanism designed to
entirely replace the repetitive code sequence in each main() func
and most of the _test.go files that have to startup (and shutdown)
the various packages they are dependent upon... and do so in the
proper order. Maintenance of these dependencies and sequences will
eventually be totally automated once each such package implements
an init() func to register their callbacks and the main() (and
test) func's replace their package initialization/shutdown sequences
with a simple set of calls to the transitions package.

At this point, all packages are integraded. What remains are the
following programs:

proxyfsd/proxyfsd
pfsalived
pfsworkout
pfs-crash
inodeworkout
fsworkout