Used etcd integration test framework as basis for our unit tests
Use etcd integration framework as basis for our unit tests.  We no
longer need a running 3 node etcd cluster for the unit tests to pass.

Add flag to Struct signifying that we are running a unit test and
therefore set different arguments to vg_up_down script.

Change vg_up_down script to take different arguments mockUp and
mockDown.  In the future we will use these to simulate in a unit
test a failed up or failed down.

Change the API to be more Go like.`

Add a couple of functions in testutils for writing unit tests.  More
to come.