Bug fixes.
Fix unpackKeyValue().  It turns that a KeyValue in an event can
have an empty value if the the event was for a key that was deleted.
Change unpackKeyValues() to notice when CreateVersion == 0 and return
a comparison function for the key that checks CreateVersion == 0 if
somebody wants to verify that key hasn't changed since it was deleted
(though this will not match correctly if the key is created and deleted
again before the check is made).

Fix a bug in unpackVgInfo() if is passed a VG that has been deleted
(probably only possible if the information is coming from an event).
In that case, don't try to unpack the (empty) value and just leave
the rest of the VgInfo with zero values.

Make up/down arguemnt to callUpDownScript() a real type, and map to
strings down where the script is called.