proxyfsd is now alive !!!
At this point, all regularly used in development and deployment are alive.

Unfortunately, `make c-test` currently fails. This is the same as:

  cdpfs
  cd jrpcclient
  ./regression_test.py --just-test-libs

Oddly, running the tests manually works... these are the steps:

Term1:
	cdpfs
	make fmt install stringer generate test python-test vet c-clean c-build c-install
	ramswift ramswift/macramswift0.conf SwiftClient.NoAuthTCPPort=4592
Term2:
	cdpfs
	cleanproxyfs proxyfsd/saioproxyfsd0.conf Logging.LogFilePath= SwiftClient.NoAuthTCPPort=4592
	mkproxyfs -N CommonVolume proxyfsd/saioproxyfsd0.conf Volume:CommonVolume.ReplayLogFileName=/tmp/CommonVolume.rlog Logging.LogFilePath= SwiftClient.NoAuthTCPPort=4592
	proxyfsd proxyfsd/saioproxyfsd0.conf Volume:CommonVolume.ReplayLogFileName=/tmp/CommonVolume.rlog Logging.LogFilePath= SwiftClient.NoAuthTCPPort=4592 JSONRPCServer.TCPPort=12347 JSONRPCServer.FastTCPPort=32347 HTTPServer.TCPPort=15347
Term3:
	cdpfs
	cd jrpcclient
	./test -r 127.0.0.1:12347/32347

The issue might have something to do with AutoFormat settings & mkproxyfs.

Meanwhile, this checking includes a line in inode/config.go that `go fmt`
has varying opinions on. For v1.10, a line in a struct initializer wants
the value to be one space to the right of the colon. For v1.11, the value
is to be aligned with the other two values for the struct. This version
captures the v1.11-preferred format and, hence, a `go fmt` will report
that inode/config.go needed to be reformatted if Go is at v1.10.